pyproject.toml
[project]
name = "instalive-live-app"
version = "0.1.0"
description = ""
authors = [
    {name = "robiulsunnyemon",email = "robiulsunyemon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "motor (==3.6.0)",
    "beanie (>=1.25.0,<=2.0.1)",
    "livekit-api (>=1.1.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "aiosmtplib (>=5.0.0,<6.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "stripe (>=11.3.0,<12.0.0)",
    "sendgrid (>=6.11.0,<7.0.0)",
]

[tool.poetry]
packages = [{include = "instalive_live_app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
